only author can edit

+++ lit_review/reviews/views.py
@@ -1,5 +1,6 @@
 from itertools import chain
 
+from django.http import Http404
 from django.views.generic import View
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required, permission_required
@@ -42,7 +43,9 @@ class TicketEdit(View):
     template_name = 'reviews/edit_ticket.html'
 
     def get(self, request, ticket_id):
-        get_object_or_404(models.Ticket, id=ticket_id)
+        ticket = get_object_or_404(models.Ticket, id=ticket_id)
+        if ticket.user != request.user:
+            raise Http404('Vous n\'avez pas la permission de modifier ce contenu.')
         edit_form = forms.TicketForm()
         delete_form = forms.DeleteTicketForm()
         return render(request, self.template_name,
@@ -123,7 +126,9 @@ class ReviewEdit(View):
     template_name = 'reviews/edit_review.html'
 
     def get(self, request, review_id):
-        get_object_or_404(models.Review, id=review_id)
+        review = get_object_or_404(models.Review, id=review_id)
+        if review.user != request.user:
+            raise Http404('Vous n\'avez pas la permission de modifier ce contenu.')
         edit_form = forms.ReviewForm()
         delete_form = forms.DeleteReviewForm()
         return render(request, self.template_name,
